Reorganize Animation into multiple files
